+++ src/Completion/CompletionInterface.php
@@ -7,7 +7,7 @@ namespace ByCerfrance\LlmApiLib\Completion;
 use ByCerfrance\LlmApiLib\Completion\Message\MessageInterface;
 use ByCerfrance\LlmApiLib\Completion\Message\RoleEnum;
 use ByCerfrance\LlmApiLib\Completion\ResponseFormat\ResponseFormatInterface;
-use ByCerfrance\LlmApiLib\Completion\Tool\ToolCollection;
+use ByCerfrance\LlmApiLib\Completion\Tool\ToolCollectionInterface;
 use ByCerfrance\LlmApiLib\Completion\Tool\ToolInterface;
 use ByCerfrance\LlmApiLib\Model\CapabilityRequirement;
 use ByCerfrance\LlmApiLib\Model\ModelInfo;
@@ -21,18 +21,18 @@ interface CompletionInterface extends CapabilityRequirement, JsonSerializable, C
     /**
      * Get tools.
      *
-     * @return ToolCollection|null
+     * @return ToolCollectionInterface|null
      */
-    public function getTools(): ?ToolCollection;
+    public function getTools(): ?ToolCollectionInterface;
 
     /**
      * With tools.
      *
-     * @param ToolCollection|ToolInterface|null ...$tools
+     * @param ToolCollectionInterface|ToolInterface|null ...$tools
      *
      * @return CompletionInterface
      */
-    public function withTools(ToolCollection|ToolInterface|null ...$tools): CompletionInterface;
+    public function withTools(ToolCollectionInterface|ToolInterface|null ...$tools): CompletionInterface;
 
     /**
      * Get max tool iterations.